requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0

